move download_database and upload_database to .data.api & add class name of a model to its hparams
