+++ mmk/data/__init__.py
@@ -1,4 +1,4 @@
-from .api import Database, FeatureProxy
+from .api import Database, FeatureProxy, download_database, upload_database
 from .data_object import DataObject
 from .factory import make_root_db, AudioFileWalker
 from .metadata import Metadata
